Add and update Bank fucntions
Added a function for added to the player bank. This includes an array
data structure which holds the values of the coins that will be
colleceted in the game. Each coin in assigned a coeficient corresponding
to the array index ('coinValue') which can be modified in GameState
initialization. Added and formatted comments.