command parts parsing update
user commands which are passed to CommandFactory with more than two
words will now recognize every word after the second as an extension of
the noun. The single letter or verb command funcutionality is retained
by giving the noun variable the value of empty string if the command
passed to the CommandFactory is only on part(string/word/character)
long. Take command is retested on coin and normal objects.  Events
tested too.