Random Object Generation Fucntion Started
Room constructor now reads for one word permission descriptions in
between the angle brackets "<>" on the Room 'Contents: ' line when
reading the .bork file. The string is separated into two parts. The
first part is split again into a String array and passed to another
helper sub routine setRandPermissions() method which then determines
which permissions should be unlocked for this room. At this time the
only key words which act as permission key words are coin and mob.
Naturally the load file was updated with a few test functions. The file
loads properly and the booleans are correctly set.